% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab2word.R
\name{tab2word}
\alias{tab2word}
\title{tab2word}
\usage{
tab2word(table, file, footer, colwid = 1, headnames = NA)
}
\arguments{
\item{table}{a table}

\item{file}{output file path}

\item{footer}{footer text}

\item{colwid}{column width}

\item{headnames}{list of strings for header column titles}
}
\value{

}
\description{
This function accepts a table made in tib2tab and outputs a formatted word document
}
\examples{
\dontrun{
tib <- dat2tib(data = cdisc_data, 
               model = lm(aval ~ arm + age + sex), 
               outcome = aval, 
               trt = arm, 
               nest = param, 
               tran='none')
               
tab <- tib2tab(mtib = tib, 
        stat_fmt = "{mean} ({sd})", 
        comp_fmt = "{estimate} ({lower_CL}, {upper_CL})")
        
tab2word(tab, 
         file = 'tabletest.docx', 
         footer = 'Table created with dat2stat R package', 
         colwid = 1)
}

}
