% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tib2tab.R
\name{tib2tab}
\alias{tib2tab}
\title{tib2tab}
\usage{
tib2tab(mtib, stat_fmt, comp_fmt, adjust = F)
}
\arguments{
\item{mtib}{A master tibble}

\item{stat_fmt}{Format for summary statistics by arm in glue syntax}

\item{comp_fmt}{Format for pairwise comparison statistics in glue syntax}

\item{adjust}{a logical value indicating whether summary statistics should be adjusted or unadjusted by model covariates}
}
\value{

}
\description{
This function accepts a master tibble made in dat2tib and creates a summary table with one row per outcome variable 
and summary statistics by treatment arms.
}
\examples{
\dontrun{
tib <- dat2tib(data = cdisc_data, 
               model = lm(aval ~ arm + age + sex), 
               outcome = aval, 
               trt = arm, 
               nest = param, 
               tran='none')
               
tib2tab(mtib = tib, 
        stat_fmt = "{mean} ({sd})", 
        comp_fmt = "{estimate} ({lower_CL}, {upper_CL})")
}

}
